Check global setting before using.

+++ plugins/quant/uninstall.php
@@ -9,6 +9,12 @@ if (!defined('WP_UNINSTALL_PLUGIN')) {
 
 // Clean up options data.
 delete_option("quant_internal_token");
-delete_option(QUANT_SETTINGS_KEY);
-delete_option(QUANT_CRON_SETTINGS_KEY);
-delete_option(QUANT_SEED_KEY);
+if (defined('QUANT_SETTINGS_KEY')) {
+    delete_option(QUANT_SETTINGS_KEY);
+}
+if (defined('QUANT_CRON_SETTINGS_KEY')) {
+    delete_option(QUANT_CRON_SETTINGS_KEY);
+}
+if (defined('QUANT_SEED_KEY')) {
+    delete_option(QUANT_SEED_KEY);
+}
